+++ GoDiveMVP/Views/Components/ProfileSheetPresentation.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+
+extension View {
+    func profileEditSheetPresentation() -> some View {
+        presentationDetents([.medium, .large])
+            .presentationDragIndicator(.visible)
+            .appSheetContentTopSpacing()
+    }
+}